# api/impl_tmp/openapi_server/models/audit_trail.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class AuditTrail(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, audit_id=None, user_id=None, healthcare_provider_id=None, is_web_access=None, is_api_access=None, access_date=None):  # noqa: E501
        """AuditTrail - a model defined in OpenAPI

        :param audit_id: The audit_id of this AuditTrail.  # noqa: E501
        :type audit_id: str
        :param user_id: The user_id of this AuditTrail.  # noqa: E501
        :type user_id: str
        :param healthcare_provider_id: The healthcare_provider_id of this AuditTrail.  # noqa: E501
        :type healthcare_provider_id: str
        :param is_web_access: The is_web_access of this AuditTrail.  # noqa: E501
        :type is_web_access: bool
        :param is_api_access: The is_api_access of this AuditTrail.  # noqa: E501
        :type is_api_access: bool
        :param access_date: The access_date of this AuditTrail.  # noqa: E501
        :type access_date: date
        """
        self.openapi_types = {
            'audit_id': str,
            'user_id': str,
            'healthcare_provider_id': str,
            'is_web_access': bool,
            'is_api_access': bool,
            'access_date': date
        }

        self.attribute_map = {
            'audit_id': 'audit_id',
            'user_id': 'user_id',
            'healthcare_provider_id': 'healthcare_provider_id',
            'is_web_access': 'is_web_access',
            'is_api_access': 'is_api_access',
            'access_date': 'access_date'
        }

        self._audit_id = audit_id
        self._user_id = user_id
        self._healthcare_provider_id = healthcare_provider_id
        self._is_web_access = is_web_access
        self._is_api_access = is_api_access
        self._access_date = access_date

    @classmethod
    def from_dict(cls, dikt) -> 'AuditTrail':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The AuditTrail of this AuditTrail.  # noqa: E501
        :rtype: AuditTrail
        """
        return util.deserialize_model(dikt, cls)

    @property
    def audit_id(self):
        """Gets the audit_id of this AuditTrail.


        :return: The audit_id of this AuditTrail.
        :rtype: str
        """
        return self._audit_id

    @audit_id.setter
    def audit_id(self, audit_id):
        """Sets the audit_id of this AuditTrail.


        :param audit_id: The audit_id of this AuditTrail.
        :type audit_id: str
        """

        self._audit_id = audit_id

    @property
    def user_id(self):
        """Gets the user_id of this AuditTrail.


        :return: The user_id of this AuditTrail.
        :rtype: str
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id):
        """Sets the user_id of this AuditTrail.


        :param user_id: The user_id of this AuditTrail.
        :type user_id: str
        """

        self._user_id = user_id

    @property
    def healthcare_provider_id(self):
        """Gets the healthcare_provider_id of this AuditTrail.


        :return: The healthcare_provider_id of this AuditTrail.
        :rtype: str
        """
        return self._healthcare_provider_id

    @healthcare_provider_id.setter
    def healthcare_provider_id(self, healthcare_provider_id):
        """Sets the healthcare_provider_id of this AuditTrail.


        :param healthcare_provider_id: The healthcare_provider_id of this AuditTrail.
        :type healthcare_provider_id: str
        """

        self._healthcare_provider_id = healthcare_provider_id

    @property
    def is_web_access(self):
        """Gets the is_web_access of this AuditTrail.


        :return: The is_web_access of this AuditTrail.
        :rtype: bool
        """
        return self._is_web_access

    @is_web_access.setter
    def is_web_access(self, is_web_access):
        """Sets the is_web_access of this AuditTrail.


        :param is_web_access: The is_web_access of this AuditTrail.
        :type is_web_access: bool
        """

        self._is_web_access = is_web_access

    @property
    def is_api_access(self):
        """Gets the is_api_access of this AuditTrail.


        :return: The is_api_access of this AuditTrail.
        :rtype: bool
        """
        return self._is_api_access

    @is_api_access.setter
    def is_api_access(self, is_api_access):
        """Sets the is_api_access of this AuditTrail.


        :param is_api_access: The is_api_access of this AuditTrail.
        :type is_api_access: bool
        """

        self._is_api_access = is_api_access

    @property
    def access_date(self):
        """Gets the access_date of this AuditTrail.


        :return: The access_date of this AuditTrail.
        :rtype: date
        """
        return self._access_date

    @access_date.setter
    def access_date(self, access_date):
        """Sets the access_date of this AuditTrail.


        :param access_date: The access_date of this AuditTrail.
        :type access_date: date
        """

        self._access_date = access_date