# api/impl_tmp/openapi_server/models/dietary_regimen.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class DietaryRegimen(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, dietary_regimen_id=None, start_date=None, end_date=None):  # noqa: E501
        """DietaryRegimen - a model defined in OpenAPI

        :param dietary_regimen_id: The dietary_regimen_id of this DietaryRegimen.  # noqa: E501
        :type dietary_regimen_id: str
        :param start_date: The start_date of this DietaryRegimen.  # noqa: E501
        :type start_date: date
        :param end_date: The end_date of this DietaryRegimen.  # noqa: E501
        :type end_date: date
        """
        self.openapi_types = {
            'dietary_regimen_id': str,
            'start_date': date,
            'end_date': date
        }

        self.attribute_map = {
            'dietary_regimen_id': 'dietary_regimen_id',
            'start_date': 'start_date',
            'end_date': 'end_date'
        }

        self._dietary_regimen_id = dietary_regimen_id
        self._start_date = start_date
        self._end_date = end_date

    @classmethod
    def from_dict(cls, dikt) -> 'DietaryRegimen':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The DietaryRegimen of this DietaryRegimen.  # noqa: E501
        :rtype: DietaryRegimen
        """
        return util.deserialize_model(dikt, cls)

    @property
    def dietary_regimen_id(self):
        """Gets the dietary_regimen_id of this DietaryRegimen.


        :return: The dietary_regimen_id of this DietaryRegimen.
        :rtype: str
        """
        return self._dietary_regimen_id

    @dietary_regimen_id.setter
    def dietary_regimen_id(self, dietary_regimen_id):
        """Sets the dietary_regimen_id of this DietaryRegimen.


        :param dietary_regimen_id: The dietary_regimen_id of this DietaryRegimen.
        :type dietary_regimen_id: str
        """

        self._dietary_regimen_id = dietary_regimen_id

    @property
    def start_date(self):
        """Gets the start_date of this DietaryRegimen.


        :return: The start_date of this DietaryRegimen.
        :rtype: date
        """
        return self._start_date

    @start_date.setter
    def start_date(self, start_date):
        """Sets the start_date of this DietaryRegimen.


        :param start_date: The start_date of this DietaryRegimen.
        :type start_date: date
        """

        self._start_date = start_date

    @property
    def end_date(self):
        """Gets the end_date of this DietaryRegimen.


        :return: The end_date of this DietaryRegimen.
        :rtype: date
        """
        return self._end_date

    @end_date.setter
    def end_date(self, end_date):
        """Sets the end_date of this DietaryRegimen.


        :param end_date: The end_date of this DietaryRegimen.
        :type end_date: date
        """

        self._end_date = end_date